Major refactoring to enable single pass codegen
This is needed to get nested class names to match how it previously
worked.

The most significant part is the introduction of
SpecExtractorHelper.createSpec which uses new functionality in nryaml
to get away from the juggling of nested Map structures and lots of casts
when dealing with parsed yaml data.

It also gets rid of VavrHelpers that are no longer needed.